refactor(int-1066): changing scripts to TS updating css to not use important changing loops from while to for
