+++ src/components/Datepicker/utils.ts
@@ -1,0 +1,14 @@
+/**
+ * @type {Array<String>}
+ */
+export const datepickerOptions: Array<string> = ['datetime', 'date']
+
+/**
+ * @type {Object}
+ */
+export const INTERNAL_VIEWS: object = {
+  CALENDAR: 'CALENDAR',
+  TIME: 'TIME',
+  YEAR: 'YEAR',
+  MONTH: 'MONTH',
+}
