+++ src/utils/index.js
@@ -7,3 +7,4 @@ export * from './number'
 export * from './objects'
 export * from './strings'
 export * from './vdom'
+export * from './timezones'
